requirements.txt
dj_database_url==2.3.0
Django==5.1.4
django_crispy_forms==2.3
django_versatileimagefield==3.1
djangorestframework_simplejwt==5.4.0
pytz==2025.1
drf-yasg==1.21.8
djangorestframework==3.15.2
django-filter==25.1
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-configurations==2.5.1
requests==2.32.3
django-cors-headers == 4.7.0
